migrating WithUserInfo extension method out of EventViewModelExtension class and inserting it into the only code that uses it, ShowEventQueryHandler as a private method
